--- a/.gitignore
+++ b/.gitignore
@@ -138,3 +138,4 @@ AI_MODEL_SYNTHESIS.md
 main
 *.tmp
 .artifacts/
+internal/cli/testdata/