--- a/.gitignore
+++ b/.gitignore
@@ -139,3 +139,8 @@ main
 *.tmp
 .artifacts/
 internal/cli/testdata/
+
+# Task and TODO temporary files
+task.md
+TODO-*.md
+TASK-*.md