… .pre-commit-config.yaml
… CLAUDE.md
… README.md
… TODO.md
… cmd/thinktank/cli_all_models_test.go
… docs/STRUCTURED_LOGGING.md
… docs/TROUBLESHOOTING.md
… docs/operations/PERFORMANCE_METRICS.md
… go.sum
… internal/cli/audit_logging_test.go
… internal/cli/help_test.go
… internal/cli/main_additional_test.go
… internal/cli/multiple_paths_integration_test.go
… internal/cli/safety_margin_integration_test.go
… internal/cli/safety_margin_test.go
… internal/cli/select_models_test.go
… internal/cli/simple_parser.go
… internal/cli/simple_parser_multiple_paths_test.go
… internal/cli/simple_parser_test.go
… internal/docs/doc_validation_test.go
… internal/integration/boundary_synthesis_flow_new_test.go
… internal/integration/boundary_synthesis_flow_test.go
… internal/integration/boundary_test_adapter.go
… internal/integration/comprehensive_e2e_workflow_test.go
… internal/integration/correlation_id_test.go
… internal/integration/invalid_synthesis_model_test.go
… internal/integration/multi_model_reliability_test.go
… internal/integration/no_synthesis_flow_test.go
… internal/integration/synthesis_with_failures_test.go
… internal/logutil/console_writer_status.go